requirements.txt
fastapi~=0.78.0
uvicorn~=0.18.1
PyJWT~=2.4.0
python-decouple
pydantic~=1.9.1
python-multipart~=0.0.5
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4